更新表模型

+++ internal/model/banners.gen.go
@@ -14,12 +14,12 @@ import (
 type Banners struct {
 	bun.BaseModel `bun:"table:banners,alias:ba"`
 
-	ID    uint64 `bun:"id,pk,autoincrement" json:"id"`
+	ID    uint32 `bun:"id,pk,autoincrement" json:"id"`
 	Title string `bun:"title,notnull" json:"title"`
 	Cover string `bun:"cover,notnull" json:"cover"`
 	Link  string `bun:"link,notnull" json:"link"`
 	// Relation: video_id -> Videos(ID)
-	VideoID uint64 `bun:"video_id,notnull" json:"video_id"`
+	VideoID uint32 `bun:"video_id,notnull" json:"video_id"`
 	Sort    uint32 `bun:"sort,notnull" json:"sort"`
 	// 1启用 0禁用
 	Status    uint8     `bun:"status,notnull" json:"status"`
