+++ internal/model/online_sessions.gen.go
@@ -12,7 +12,7 @@ import (
 type OnlineSessions struct {
 	bun.BaseModel `bun:"table:online_sessions,alias:os"`
 
-	ID           uint64    `bun:"id,pk,autoincrement" json:"id"`
+	ID           uint32    `bun:"id,pk,autoincrement" json:"id"`
 	SessionKey   string    `bun:"session_key,notnull,unique" json:"session_key"`
 	IP           string    `bun:"ip,notnull" json:"ip"`
 	LastActiveAt time.Time `bun:"last_active_at,notnull" json:"last_active_at"`
