+++ internal/model/user_favorites.gen.go
@@ -14,11 +14,11 @@ import (
 type UserFavorites struct {
 	bun.BaseModel `bun:"table:user_favorites,alias:uf"`
 
-	ID uint64 `bun:"id,pk,autoincrement" json:"id"`
+	ID uint32 `bun:"id,pk,autoincrement" json:"id"`
 	// Relation: user_id -> Users(ID)
-	UserID uint64 `bun:"user_id,notnull,unique:uk_user_video" json:"user_id"`
+	UserID uint32 `bun:"user_id,notnull,unique:uk_user_video" json:"user_id"`
 	// Relation: video_id -> Videos(ID)
-	VideoID   uint64    `bun:"video_id,notnull,unique:uk_user_video" json:"video_id"`
+	VideoID   uint32    `bun:"video_id,notnull,unique:uk_user_video" json:"video_id"`
 	CreatedAt time.Time `bun:"created_at,notnull" json:"created_at"`
 	User      *Users    `bun:"rel:belongs-to,join:user_id=id" json:"-"`
 	Video     *Videos   `bun:"rel:belongs-to,join:video_id=id" json:"-"`
