+++ internal/model/user_login_logs.gen.go
@@ -14,9 +14,9 @@ import (
 type UserLoginLogs struct {
 	bun.BaseModel `bun:"table:user_login_logs,alias:ull"`
 
-	ID uint64 `bun:"id,pk,autoincrement" json:"id"`
+	ID uint32 `bun:"id,pk,autoincrement" json:"id"`
 	// Relation: user_id -> Users(ID)
-	UserID    uint64 `bun:"user_id,notnull" json:"user_id"`
+	UserID    uint32 `bun:"user_id,notnull" json:"user_id"`
 	Username  string `bun:"username,notnull" json:"username"`
 	IP        string `bun:"ip,notnull" json:"ip"`
 	UserAgent string `bun:"user_agent,notnull" json:"user_agent"`
