+++ internal/model/user_play_history.gen.go
@@ -14,15 +14,15 @@ import (
 type UserPlayHistory struct {
 	bun.BaseModel `bun:"table:user_play_history,alias:uph"`
 
-	ID uint64 `bun:"id,pk,autoincrement" json:"id"`
+	ID uint32 `bun:"id,pk,autoincrement" json:"id"`
 	// Relation: user_id -> Users(ID)
-	UserID uint64 `bun:"user_id,notnull,unique:uk_user_video" json:"user_id"`
+	UserID uint32 `bun:"user_id,notnull,unique:uk_user_video" json:"user_id"`
 	// Relation: video_id -> Videos(ID)
-	VideoID uint64 `bun:"video_id,notnull,unique:uk_user_video" json:"video_id"`
+	VideoID uint32 `bun:"video_id,notnull,unique:uk_user_video" json:"video_id"`
 	// Relation: play_source_id -> PlaySources(ID)
-	PlaySourceID uint64 `bun:"play_source_id,notnull" json:"play_source_id"`
+	PlaySourceID uint32 `bun:"play_source_id,notnull" json:"play_source_id"`
 	// Relation: episode_id -> PlayEpisodes(ID)
-	EpisodeID uint64 `bun:"episode_id,notnull" json:"episode_id"`
+	EpisodeID uint32 `bun:"episode_id,notnull" json:"episode_id"`
 	// 播放进度（秒）
 	Progress uint32 `bun:"progress,notnull" json:"progress"`
 	// 总时长（秒）
